requirements-http.txt
# HTTP MCP Server Dependencies
# Core MCP and web framework dependencies
fastmcp>=0.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Existing Proxmox dependencies
mcp>=1.0.0
paramiko>=3.0.0
proxmoxer>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
urllib3>=2.0.0

# Additional HTTP server dependencies
pydantic>=2.5.0
starlette>=0.27.0
anyio>=4.0.0

# Development and production dependencies
uvloop>=0.19.0  # For better async performance
gunicorn>=21.2.0  # For production WSGI deployment